Changes to file structure; reading in of binary file for model
I've changed some file structure in the project, but a lot of it needs
to be cleaned up still.  The bigger update is that meshes are loaded
once from file, and are loaded from a binary format that another project
exports, given relatively simple .obj files.  This project should
display a spinning dragon on screen with 100000 faces and shouldn't take
more than ~15s to load.  Further optimizations are possible.